refactor: modularize calendar into components and composables
